Adding `lib` support to crossgen (#12)
* [WIP] Adding `lib` support to crossgen

So, this is just quickly something I threw together on the plane – looking for
some feedback!

The CLI now has a `--lib` flag which is taken into account when creating
a template. `Template::new` still exists, but is deprecated and points
to `Template::gen_bin` (just in case this shouldn't be a breaking change).

Actually generating the templates I'm not 100% sure what differences
might be wanted/required for bin vs lib. So I just copied the templates
directory for now.

Also the write block is quite ugly now. Considering that this approach
generally isn't very modular, I'm wondering if it wouldn't be possible
to have a folder with travis "snippets" that are then assembled according
to some user config (e.g. I want android but not wasm, but it's also a binary?)

Again...WIP – happy to change stuff!

Cheers!

Signed-off-by: Katharina Fey <kookie@spacekookie.de>

* Making the `lib` flag a bool, removing the explicit default

Signed-off-by: Katharina Fey <kookie@spacekookie.de>

* Running `rustfmt` for the PR

Signed-off-by: Katharina Fey <kookie@spacekookie.de>