Read Cargo.toml from target dir
When running `crossgen` on a directory with a missing `Cargo.toml` it
would panic with an error message.

Debugging what happend, I've found it failed to find a `Cargo.toml` file, as it don't exist.
To my surprise, when I've cloned the repo and tried to reproduce using `cargo run -- ~/path/to/my-dir`, there was no panic!

This happened because the `Cargo.toml` was found on `crossgen` repo, not on my targe repo, as expected.

This commit changes the location of the manifest file to use the directory parsed from the arguments.

Signed-off-by: Bruno Tavares <connect+github@bltavares.com>