Define the type of artifact to be produced by cargo.
s command `cargo` subcommand is exclusively intended to be used to help
with [working arround](https://github.com/rust-lang/rust/issues/51009)
how crate types are defined, in order to help with cross-platform
builds.

It is currently [not possible](https://github.com/rust-lang/cargo/issues/6160) to define a
single `crate-type` override on `cargo build`, which causes libs
intended to be used on other languages to compile more than one type of
crate.

This commit adds a dependency on this [new CLI app](https://github.com/bltavares/cargo-crate-type) to be able to workaround the limitations of Cargo.

We are now able to enable the `-C lto` optimization again, as we will
only compile a single type of artifact

Signed-off-by: Bruno Tavares <connect+github@bltavares.com>